# Pre-process ClusterGroups' Selector on Spec changes