--- a/internal/cmd/controller/controllers/clustergroup/controller.go
+++ b/internal/cmd/controller/controllers/clustergroup/controller.go
@@ -3,7 +3,6 @@ package clustergroup
 
 import (
 	"context"
-	"k8s.io/apimachinery/pkg/runtime"
 	"sort"
 
 	"github.com/sirupsen/logrus"
@@ -15,6 +14,7 @@ import (
 	"github.com/rancher/wrangler/pkg/kv"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/runtime"
 )
 
 type Store interface {