<!-- Specify the issue ID that this pull request is solving -->
Refers to _ToBeCreated_
<!-- Make sure that the referenced issue provides steps to reproduce it -->

We have profiling data that shows how calculating if a Fleet Cluster belongs to a ClusterGroup can cause a considerable use of CPU under certain scenarios. In particular, most of this usage is dedicated to processing the field `.Spec.Selector` from each ClusterGroup, as shown in this screenshot:

![Screenshot 2023-12-14 at 13 18 52](https://github.com/rancher/fleet/assets/4057165/93429232-8b15-42ee-b496-fddeb6e72017)

The changes in this PR add some sort of caching mechanism for this processing, making the `clustergroup` controller to store the relevant object every time a change in the `ClusterGroup` spec is observed. Then the targeting mechanism can reuse this pre-processed selector in a more efficient way.

TODO
------
- [ ] Add unit tests for `Store`