@manno 2023-12-14
I quickly added a benchmark test for this here: https://github.com/rancher/fleet/blob/clustegroup-store-0.8-bench/internal/cmd/controller/target/selector_test.go#L19

Amazing that this selector shows up in the profiling, from my quick measurement:

Before:

```
% go test -bench=. ./internal/cmd/controller/target                                                                                                                                                                                                    
goos: darwin
goarch: arm64
pkg: github.com/rancher/fleet/internal/cmd/controller/target
BenchmarkStore-10    	   10000	   1801391 ns/op
PASS
ok  	github.com/rancher/fleet/internal/cmd/controller/target	18.311s
go test -bench=. ./internal/cmd/controller/target  132,37s user 10,70s system 687% cpu 20,806 total
```

After:
```
% go test -bench=. ./internal/cmd/controller/target
goos: darwin
goarch: arm64
pkg: github.com/rancher/fleet/internal/cmd/controller/target
BenchmarkStore-10    	   10000	   1379797 ns/op
PASS
ok  	github.com/rancher/fleet/internal/cmd/controller/target	14.095s
go test -bench=. ./internal/cmd/controller/target  118,32s user 1,60s system 724% cpu 16,563 total
```