@raulcabello 2024-01-10
Looks good to me. Should we target v0.9 instead of v0.8? just for avoiding the risk of introducing a bug in the stable release.